Lock CSV package dependency to working interface
The node-csv-parser interface has changed drastically over the last few years including the interface. So locking down to version 0.1.0 of the csv package brings in the last known working interface that this package expects. And this allows the build.sh and prepare-data.js scripts to run correctly (after npm install is run).